+++ tests/testthat/test-draw-panel.R
@@ -238,3 +238,15 @@ test_that("drop x label", {
   p <- arphitgg(foo, agg_aes(x=date,y=y),layout="2v")+agg_line(panel=c("1","2"))
   expect_true(check_graph(p,  "draw-panel-drop-first-ts-auto-not-required"))
 })
+
+## Formatting of y labels ==================
+
+test_that("Formatting y labels", {
+  #Many decimal points
+  p <- arphitgg()+agg_ylim(0,3e-6,4)
+  expect_true(check_graph(p, "draw-panel-ylabel-format-decimals"))
+
+  # Thousands separator #359
+  p <- arphitgg()+agg_ylim(0,30000,4)
+  expect_true(check_graph(p, "draw-panel-ylabel-format-thousands"))
+})